+++ client/src/components/tags/manage/EditTag.vue
@@ -10,24 +10,54 @@
     }"
     @update:visible="updateVisible"
   >
-    <form @submit.prevent="onSubmit">
+    <form @submit="save">
       <div class="field">
         <label for="name">Name</label>
         <PrimeInputText
           id="name"
+          v-model.trim="name.value.value"
           type="text"
+          :disabled="isFormDisabled"
+          :invalid="name.errors.value.length > 0"
         />
+        <small
+          v-if="name.errors.value.length > 0"
+          class="p-error"
+        >
+          <span v-if="name.errors.value.length === 1">{{ name.errorMessage.value }}</span>
+          <span
+            v-for="error in name.errors.value"
+            :key="error"
+            >{{ error }} <br
+          /></span>
+        </small>
       </div>
 
       <div class="footer">
         <div class="result">
-          <p>Result can have a lot of content and we should wrap it correctly</p>
+          <PrimeInlineMessage
+            v-if="mutation.isSuccess.value"
+            severity="success"
+            >Tag saved</PrimeInlineMessage
+          >
+          <ApiError
+            v-if="mutation.isError.value"
+            :error="mutation.error.value"
+          >
+            <template #message>
+              <p>
+                Saving the tag failed, please refer to the error information below for more details.
+              </p>
+            </template>
+          </ApiError>
         </div>
         <div class="actions">
           <PrimeButton
             type="submit"
             label="Save"
             icon="pi pi-save"
+            :disabled="isFormDisabled"
+            :loading="mutation.isPending.value"
           />
         </div>
       </div>
@@ -40,6 +70,13 @@ import { computed } from 'vue';
 import PrimeDialog from 'primevue/dialog';
 import PrimeInputText from 'primevue/inputtext';
 import PrimeButton from 'primevue/button';
+import PrimeInlineMessage from 'primevue/inlinemessage';
+import { useField, useForm } from 'vee-validate';
+import { toTypedSchema } from '@vee-validate/yup';
+import * as yup from 'yup';
+import { useChoresApiUpsertTag } from '@/composables/queries/chores-api.js';
+import { useQueryClient } from '@tanstack/vue-query';
+import ApiError from '@/components/common/ApiError.vue';
 
 const props = defineProps({
   tag: {
@@ -50,19 +87,55 @@ const props = defineProps({
 const emit = defineEmits(['close']);
 const isEdit = computed(() => props.tag != null);
 
+//#region form
+
+const { handleSubmit, meta } = useForm({
+  validationSchema: toTypedSchema(
+    yup.object({
+      name: yup.string().required().label('Name'),
+    }),
+  ),
+  initialValues: {
+    name: props.tag?.name,
+  },
+});
+
+const isFormDisabled = computed(
+  () => mutation.isPending.value === true || mutation.isSuccess.value === true,
+);
+
+const name = useField('name');
+
+//#endregion
+
 //#region create/update
 
-const onSubmit = () => {
-  console.log('submit');
-};
+const queryClient = useQueryClient();
+const mutation = useChoresApiUpsertTag(queryClient);
+
+const save = handleSubmit.withControlled(async (values) => {
+  try {
+    const payload = {
+      name: values.name,
+    };
+    if (isEdit.value === true) payload.id = props.tag.id;
+    await mutation.mutateAsync(payload);
+  } catch (e) {
+    console.error(e);
+  }
+});
 
 //#endregion
 
 const updateVisible = (value) => {
   if (value === false) tryClose();
 };
-const tryClose = () => {
-  emit('close');
+const tryClose = (force = false) => {
+  let close = true;
+  if (!force && meta.value.dirty && mutation.isSuccess.value !== true)
+    // TODO primevue confirm?
+    close = confirm('There may be unsaved changes, are you sure you want to stop editing?');
+  if (close) emit('close');
 };
 </script>
 
@@ -80,6 +153,10 @@ form {
   .footer {
     @include flex-row-justify-between-wrapping;
     flex-wrap: wrap-reverse;
+
+    .result {
+      max-width: 100%;
+    }
   }
 }
 </style>
